requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.0
google-generativeai>=0.4.1
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
